+++ CHANGELOG.md
@@ -26,6 +26,12 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/infrastructure/compare/release-4.15.1...master
 
+### Fixed
+- Spurious apt warnings have been eliminated ([JPERF-642](https://ecosystem.atlassian.net/browse/JPERF-642))
+
+# Added
+- a new, recommended method to add new repositories to Ubuntu: com.atlassian.performance.tools.infrastructure.api.os.Ubuntu.addRepository(com.atlassian.performance.tools.ssh.api.SshConnection, java.lang.String, java.lang.String)
+
 ## [4.15.1] - 2020-05-13
 [4.15.1]: https://github.com/atlassian/infrastructure/compare/release-4.15.0...release-4.15.1